Separate functions into individual modules to improve code readability and maintainability
