+++ src/modules/apiCollection.js
@@ -1,0 +1,8 @@
+const moviesURL = "https://api.tvmaze.com/shows";
+
+const involvementURL =
+  "https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/";
+
+const involvementApiKey = "KTd9w8KpncZIww7GFHXJ";
+
+export { moviesURL, involvementURL, involvementApiKey };
